refactor(server): split client.rs god-file into modules
